Adding Skills page and making changes
