+++ src/pages/Error/Maintenance.js
@@ -9,7 +9,7 @@ const Maintenance = ({name}) => {
                 <h1>Oops!! Looks like you got me at a bad time.</h1>
                 <h2> My {name} page is being currently being improved. </h2>
                 <LuConstruction size="5em" color='rgb(48, 48, 50)'/>
-                <p>Please leave you email found in the footer and I will notify you when this page is operational.</p>
+                <p>Please leave your email found in the footer and I will notify you when this page is operational.</p>
             </div>
         </>
     );
