+++ src/pages/portfolio/Portfolio.js
@@ -117,7 +117,6 @@ const Portfolio = () => {
 
     return(
         <>
-            <Analytics />
             <PortfolioGrid />
         </>
     )
